+++ PhysiCsEngine/PlaneCollider.h
@@ -7,7 +7,7 @@ class PlaneCollider : public Collider
 public:
 
 	PlaneCollider();
-	PlaneCollider( float i_a, float i_b, float i_c, float i_d );
+	PlaneCollider( double i_a, double i_b, double i_c, double i_d );
 	PlaneCollider( PCEHVector3 i_oOrthogonalVector, PCEHVector3 i_oPoint );
 	PlaneCollider( PCEHVector3 i_oFirstVector, PCEHVector3 i_oSecondVector, PCEHVector3 i_oPoint );
 
@@ -21,8 +21,8 @@ private:
 
 private:
 
-	float m_a;
-	float m_b;
-	float m_c;
-	float m_d;
+	double m_a;
+	double m_b;
+	double m_c;
+	double m_d;
 };
