+++ PhysiCsEngine/SphereCollider.h
@@ -9,19 +9,19 @@ class SphereCollider : public Collider
 public:
 
 	SphereCollider();
-	SphereCollider(float i_r, float i_x, float i_y, float i_z = 0);
+	SphereCollider(double i_r, double i_x, double i_y, double i_z = 0);
 
 	const PCEPoint getCenter() const				{return mCenter;}
 	void setCenter(const PCEPoint i_center)			{mCenter = i_center;}
-	const float getRadius() const					{return mRadius;}
-	void setRadius(const float i_radius)			{mRadius = i_radius;}
+	const double getRadius() const					{return mRadius;}
+	void setRadius(const double i_radius)			{mRadius = i_radius;}
 
 	const Collider* subShapeCollidedWith(const PCEHVector3& i_point) const;
 	const Collider* subShapeCollidedWith(const Collider& i_collider) const;
 
 private:
 	PCEPoint mCenter;
-	float mRadius;
+	double mRadius;
 
 	const Collider* SphereCollider::subShapeCollidedWith(const SphereCollider* i_sphere) const;
 };
