+++ StarPhysics/GraphicComponent.h
@@ -9,7 +9,7 @@
 class GraphicComponent : public Component
 {
 public:
-	GraphicComponent(const PCEString& i_sImagePath, PCEPoint i_oPosition, float i_fRotation, const PositionableObject * i_oOwner, bool i_bEnabled = true);
+	GraphicComponent(const PCEString& i_sImagePath, PCEPoint i_oPosition, double i_fRotation, const PositionableObject * i_oOwner, bool i_bEnabled = true);
 	~GraphicComponent(void);
 
 	void OnRender(Bitmap* renderTarget);
@@ -22,7 +22,7 @@ private:
 
 	void DestroyBitmap();
 
-	virtual void update( float i_deltaTime = 0 );
+	virtual void update( double i_deltaTime = 0 );
 
 private:
